refactor: move implementation of Prepared and Callable statements to PgPreparedStatement and PgCallableStatement
Note: the code is moved as is, no reformatting happens on the way for easier comparison
